pipelines/urban_greening_targets/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urban-greening-targets-pipeline"
version = "0.1.0"
description = "CLI pipeline to collect city urban greening targets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "requests>=2.31.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=4.9.0",
  "pdfplumber>=0.11.0",
  "pydantic>=2.5.0",
  "pandas>=2.1.0",
  "python-dotenv>=1.0.0",
  "tenacity>=8.2.0",
  "tqdm>=4.66.0",
  "curl_cffi>=0.7.0",
]

[project.scripts]
urban-greening-targets = "urban_greening_targets.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

